requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
